Add ability to know which space was clicked
Some class redesign was done, adding Game and MyScene. MyBox will handle
updating its contents. Game will handle keeping track of player turns
and holding the vector of pointers to the MyBox spaces. Game has a
function to return a pointer to the MyBox space at a particular position
on the screen. MyScene uses that function to call a mouse event on the
correct MyBox space.