Add change of image when player turn changes
A signal and slot were added to facilitate changing the player turn
image each play. In this process, a bit of refactoring went on, removing
unnecessary static functions and placing a pointer to Game in MyBox.
This required adding forward declations in both of those header files.

Also added some cleanup in the MainWindow destructor, though more needs
to be handled throughout the program.